requirements.txt
google-generativeai
load_dotenv
streamlit
